feat: support google image resolution billing

+++ backend/app/schemas/user.py
@@ -114,7 +114,7 @@ class ImageCreditRechargeRequest(BaseModel):
     """Admin image credit recharge request."""
 
     user_id: int
-    amount: int = Field(..., gt=0)
+    amount: Decimal = Field(..., gt=0)
     reason: Optional[str] = Field(None, max_length=255)
 
 
@@ -122,5 +122,5 @@ class ImageCreditDeductRequest(BaseModel):
     """Admin image credit deduct request."""
 
     user_id: int
-    amount: int = Field(..., gt=0)
+    amount: Decimal = Field(..., gt=0)
     reason: Optional[str] = Field(None, max_length=255)
